# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Claneo SMT BackendApi Documentation",
        "description": "Claneo SMT BackendApi",
        "contact": {
            "email": "mrudchenko@brightgrove.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Backend API Server"
        }
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register new user by admin",
                "description": "Accessible for only User with the 'Admin' Role",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "The given data was invalid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login to get Auth Token",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad Credentials"
                    },
                    "422": {
                        "description": "The given data was invalid"
                    },
                    "403": {
                        "description": "User should accept Privacy Policy first"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout to remove Auth Token",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Logged out"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/change-pwd": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Change password for non-logged user",
                "operationId": "change-pwd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The given data was invalid"
                    }
                }
            }
        },
        "/forgot-pwd": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send reset password link",
                "operationId": "forgot-pwd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgotPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The given data was invalid"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ChangePasswordRequest": {
                "title": "Change Password Request",
                "description": "Change Password Request",
                "properties": {
                    "email": {
                        "title": "Email",
                        "description": "Email of user",
                        "type": "string",
                        "example": "user@loc"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password of user",
                        "type": "string",
                        "example": "12345"
                    },
                    "password_confirmation": {
                        "title": "password_confirmation",
                        "description": "Enter the same password again",
                        "type": "string",
                        "example": "12345"
                    },
                    "token": {
                        "title": "Token",
                        "description": "Token from Email",
                        "type": "string",
                        "example": "noC1CHhmimTuwFpx"
                    }
                },
                "type": "object"
            },
...
            "ClientResource": {
                "title": "Client Resource",
                "description": "Client Resource",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "ID of Client",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Name of client",
                        "type": "string",
                        "example": "Client"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "created date of client",
                        "type": "string",
                        "example": "2021-10-07T19:34:40.000000Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "updated date of client",
                        "type": "string",
                        "example": "2021-10-07T19:34:40.000000Z"
                    },
                    "user": {
                        "title": "user",
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "ForgotPasswordRequest": {
                "title": "Reset Password Request",
                "description": "Reset Password Request",
                "properties": {
                    "email": {
                        "title": "Email",
                        "description": "Email of user",
                        "type": "string",
                        "example": "user@loc"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "title": "Login Request",
                "description": "Login Request",
                "properties": {
                    "email": {
                        "title": "Email",
                        "description": "Email of user",
                        "type": "string",
                        "example": "user@loc"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password of user",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "LoginResource": {
                "title": "Login Resource",
                "description": "Login Resource",
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/ClientResource"
                    },
                    "token": {
                        "title": "token",
                        "description": "Auth Token",
                        "type": "string",
                        "example": "17|1VtteyF1QBiywXwNLQiK80pq7ioJvVAq1C6C6VwO"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "title": "Register Request",
                "description": "Register Request",
                "properties": {
                    "email": {
                        "title": "Email",
                        "description": "Email of user",
                        "type": "string",
                        "example": "user@loc"
                    },
                    "first_name": {
                        "title": "first_name",
                        "description": "first_name of user",
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "title": "last_name",
                        "description": "last_name of user",
                        "type": "string",
                        "example": "Doe"
                    },
                    "roles": {
                        "title": "roles",
                        "description": "Roles of user",
                        "type": "string",
                        "example": [
                            5,
                            6
                        ]
                    }
                },
                "type": "object"
            },
...
            "RoleResource": {
                "title": "Role Resource",
                "description": "Role Resource",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "ID of Role",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": " Name of Role",
                        "type": "string",
                        "example": "SEO"
                    },
                    "description": {
                        "title": "description",
                        "description": "Description of role",
                        "type": "string",
                        "example": "SEO Manager"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "created date of user",
                        "type": "string",
                        "example": "2021-10-07T19:34:40.000000Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "updated date of user",
                        "type": "string",
                        "example": "2021-10-07T19:34:40.000000Z"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "User Resource",
                "description": "User Resource",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "ID of User",
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "title": "first_name",
                        "description": "First Name of user",
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "title": "last_name",
                        "description": "Last Name of user",
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email of user",
                        "type": "string",
                        "example": "user@loc"
                    },
                    "email_verified_at": {
                        "title": "email_verified_at",
                        "description": "Email verified date of user",
                        "type": "string",
                        "example": "NULL"
                    },
                    "privacy_policy_flag": {
                        "title": "privacy_policy_flag",
                        "description": "Has the user read the privacy policy??",
                        "type": "integer",
                        "example": 0
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "created date of user",
                        "type": "string",
                        "example": "2021-10-07T19:34:40.000000Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "updated date of user",
                        "type": "string",
                        "example": "2021-10-07T19:34:40.000000Z"
                    },
                    "roles": {
                        "title": "roles",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RoleResource"
                        },
                        "collectionFormat": "multi"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        []
    ]
}